ticket_purchase
This is one of my first projects in Python. The application is a command line based. The user is able to purchase tickets as long as the tickets aren't sold out. After 100 tickets have been sold the user will get a message saying the tickets are sold out. For every transaction there is a two dollar service fee.